backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic[email]
python-dotenv
httpx
passlib[bcrypt]
python-jose[cryptography]
pymongo
redis[asyncio]
